+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Unreleased changes
 
-[Full changelog](https://github.com/mozilla/glean.js/compare/v2.0.3...main)
+[Full changelog](https://github.com/mozilla/glean.js/compare/v2.0.4...main)
+
+# v2.0.4 (2023-10-10)
+
+[Full changelog](https://github.com/mozilla/glean.js/compare/v2.0.3...v2.0.4)
 
 * [#1772](https://github.com/mozilla/glean.js/pull/1772): Fix bug where `window.Glean` functions were getting set on non-browser properties.
 * [#1784](https://github.com/mozilla/glean.js/pull/1784): Store `window.Glean` debugging values in `sessionStorage`. This will set debug options on page init while the current session is still active.